% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_color_vector}
\alias{create_color_vector}
\title{Wrap color vector in function.}
\usage{
create_color_vector(colors = statworx_colors, ...)
}
\arguments{
\item{colors}{List of colors (either statworx_colors or statworx_colors_blue_black)}

\item{...}{Name of color from "statworx_colors"}
}
\value{
Hex code of the color
}
\description{
Wrap color vector in function.
}
