% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_statworx_palette}
\alias{create_statworx_palette}
\title{Create a color palette function.}
\usage{
create_statworx_palette(
  palette = "statworx_palette",
  reverse = FALSE,
  col_list = "",
  ...
)
}
\arguments{
\item{palette}{A color palette or "custom".}

\item{reverse}{If true, order of palette is reversed.}

\item{col_list}{Provide a vector of colors if @param palette = "custom".}
}
\value{
Palette generating function based on "colorRampPalette".
}
\description{
Create a color palette function.
}
