% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{statworx_grey1}
\alias{statworx_grey1}
\title{statworx Theme based on theme_minimal}
\usage{
statworx_grey1(font = "Arial")
}
\arguments{
\item{font}{set default font}
}
\description{
Sets a pre-defined theme as the standard ggplot theme via
\code{\link[ggplot2:ggtheme]{theme_minimal}}
}
\examples{
\dontrun{
library(statworxTheme)
statworx_grey1()
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
geom_point()
}
}
\keyword{theme}
