% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{statworx_grey2}
\alias{statworx_grey2}
\title{statworx Theme based on theme_classic}
\usage{
statworx_grey2()
}
\description{
Sets a pre-defined theme as the standard ggplot theme via
\code{\link[ggplot2:theme_get]{theme_set}} and \code{\link[ggplot2:ggtheme]{theme_classic}}
}
\examples{
\dontrun{
library(statworxTheme)
statworx_grey2()
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
geom_point()
}

}
\keyword{theme}
