% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{statworx_theme_1}
\alias{statworx_theme_1}
\title{statworx Theme based on theme_classic}
\usage{
statworx_theme_1()
}
\description{
statworx Theme based on theme_classic
}
\examples{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +   geom_point() +  statworx_theme_1()

}
\keyword{theme}
