% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{statworx_theme_2}
\alias{statworx_theme_2}
\title{statworx Theme based on theme_minimal}
\usage{
statworx_theme_2()
}
\description{
Sets a pre-defined theme as the standard ggplot theme via
\code{\link[ggplot2:theme_get]{theme_set}} and \code{\link[ggplot2:ggtheme]{theme_minimal}}
}
\examples{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
geom_point() +  statworx_theme_2()
}
\keyword{theme}
